Update skip write component config get/set name
Internally the variables are private, so name isn't a big issue there.

Internally to DataObject the method calls are `writeComponents` (called
from `write`) and `skipWriteComponents()`, which is why the get/set
functions in the EditableColumns component were named that way.

However it is a bit confusing, and the variables that are passed can be
either boolean, or a configuration array (which isn't very well
documented by core)... adding to the confusion.

By calling it a "write config" it is a bit clearer the purpose of the
function calls and what the author of consuming code intends to happen
at that point. This came as feedback on the [original PR][1]

[1]: https://github.com/symbiote/silverstripe-gridfieldextensions/pull/350